cld: Integrate qca_main changes
Includes additional logging, additional checking around type casts,
several field and local changes to avoid casts (implicit orexplicit),
compile-time asserts to validate that required explicit casts are
statically determined to be always safe, and the occasional runtime
assert for other required casts. Integrate other fixes that were
brought to qca_main perf_pwr_offload

CL Id: 1988194

+++ cld/CORE/COMMON/ol_txrx_ctrl_api.h
@@ -142,7 +142,7 @@ struct peer_ratectrl_params_t {
     u_int32_t ni_flags;
     u_int32_t ni_chwidth;
     u_int16_t ni_htcap;
-    u_int16_t ni_vhtcap;
+    u_int32_t ni_vhtcap;
     u_int16_t ni_phymode;
     u_int16_t ni_rx_vhtrates;
     u_int8_t ht_rates[MAX_SPATIAL_STREAM * 8];
